+++ system/platform-core/src/main/java/org/platformlambda/core/annotations/PreLoad.java
@@ -24,11 +24,14 @@ import java.lang.annotation.*;
  * This indicates the class is a service to be preloaded.
  * (for a class to be preloaded, it must use a default constructor without arguments)
  * <p>
- * envInstances overrides instances from multiple sources.
+ * customSerializer is optional. It allows you to apply a pre-configured ObjectMapper suitable for
+ * your user function. It should point to a fully qualified classpath to a class implementing
+ * the CustomSerializer interface.
  * <p>
- * 1. To get value from an environment variable, use this format ${ENV_VAR_NAME:defaultValue}.
- * 2. To get value from application.properties or application.yml, just set it to the parameter name.
- * Note that System property can override the application.properties/application.yml config.
+ * envInstances is optional. If present, it must be a parameter in application.properties (or application.yml).
+ * The parameter may fetch value from an environment variable using "${ENV_VAR:default_value}" format.
+ * If the parameter does not exist, or it does not resolve to a numeric value, the "instances" value in this
+ * annotation will be used instead.
  */
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
@@ -36,6 +39,7 @@ import java.lang.annotation.*;
 public @interface PreLoad {
 
     String route();
+    String customSerializer() default "";
     int instances() default 1;
     String envInstances() default "";
     boolean isPrivate() default true;