+++ system/platform-core/src/main/java/org/platformlambda/core/system/Platform.java
@@ -428,6 +428,22 @@ public class Platform {
         return service.isPrivate();
     }
 
+    public void setCustomSerializer(String route, CustomSerializer mapper) {
+        if (!hasRoute(route)) {
+            throw new IllegalArgumentException(ROUTE+route+NOT_FOUND);
+        }
+        ServiceDef service = registry.get(route);
+        if (service == null) {
+            throw new IllegalArgumentException(ROUTE+route+NOT_FOUND);
+        }
+        service.setCustomSerializer(mapper);
+        if (mapper == null) {
+            log.info("Custom serializer cleared for {}", route);
+        } else {
+            log.info("Serializer {} assigned to {}", mapper.getClass(), route);
+        }
+    }
+
     /**
      * Check if the route is trackable
      *